[feat] adgroup日志

+++ adm/src/main/java/com/tecdo/adm/delivery/service/impl/AdGroupServiceImpl.java
@@ -17,8 +17,11 @@ import com.tecdo.adm.common.cache.AdGroupCache;
 import com.tecdo.adm.delivery.service.IAdGroupService;
 import com.tecdo.adm.delivery.service.IAdService;
 import com.tecdo.adm.delivery.service.ITargetConditionService;
+import com.tecdo.adm.log.service.IBizLogApiService;
 import com.tecdo.starter.log.exception.ServiceException;
 import com.tecdo.starter.mp.entity.BaseEntity;
+import com.tecdo.starter.mp.entity.IdEntity;
+import com.tecdo.starter.mp.entity.StatusEntity;
 import com.tecdo.starter.mp.enums.BaseStatusEnum;
 import com.tecdo.starter.mp.vo.BaseVO;
 import com.tecdo.starter.tool.BigTool;
@@ -27,6 +30,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.*;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -39,6 +43,7 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
 
     private final ITargetConditionService conditionService;
     private final IAdService adService;
+    private final IBizLogApiService bizLogApiService;
 
     @Override
     public boolean add(AdGroupVO vo) {
@@ -47,12 +52,24 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
 
     @Override
     public boolean edit(AdGroupVO vo) {
-        if (vo.getId() != null && updateById(vo)) {
-            conditionService.deleteByAdGroupIds(Collections.singletonList(vo.getId()));
-            conditionService.saveBatch(vo.listCondition());
-            return true;
+        if (vo.getId() != null) {
+            logByUpdate(vo);
+            if (updateById(vo)) {
+                conditionService.deleteByAdGroupIds(Collections.singletonList(vo.getId()));
+                conditionService.saveBatch(vo.listCondition());
+                return true;
+            }
         }
         return false;
+    }
+
+    private void logByUpdate(AdGroupVO afterVO) {
+        AdGroup adGroup = getById(afterVO.getId());
+        AdGroupVO beforeVO = Objects.requireNonNull(com.tecdo.starter.tool.util.BeanUtil.copy(adGroup, AdGroupVO.class));
+        List<TargetCondition> conditions = conditionService.listCondition(afterVO.getId());
+        List<TargetConditionVO> conditionVOs = Objects.requireNonNull(com.tecdo.starter.tool.util.BeanUtil.copy(conditions, TargetConditionVO.class));
+        beforeVO.setConditionVOs(conditionVOs);
+        bizLogApiService.logByUpdateAdGroup(beforeVO, afterVO);
     }
 
     @Override
@@ -67,6 +84,7 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
     public boolean logicDelete(List<Integer> ids) {
         if (CollUtil.isEmpty(ids)) return true;
         Date date = new Date();
+        List<StatusEntity> adStatusList = adService.listStatus(ids);
         List<AdGroup> entities = ids.stream().map(id -> {
             AdGroup entity = new AdGroup();
             entity.setId(id);
@@ -77,6 +95,7 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
         updateBatchById(entities);
         List<Integer> adIds = adService.listIdByGroupIds(ids);
         adService.logicDelete(adIds);
+        bizLogApiService.logByDeleteAdGroup(ids, adStatusList);
         return true;
     }
 
@@ -134,14 +153,22 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
         if (entity == null) {
             return false;
         }
+        logByUpdateListInfo(vo, entity);
+        return updateById(entity);
+    }
+
+    private void logByUpdateListInfo(SimpleAdGroupUpdateVO vo, AdGroup entity) {
+        AdGroupVO beforeVO = Objects.requireNonNull(BeanUtil.copyProperties(entity, AdGroupVO.class));
         entity.setName(vo.getName());
         entity.setDailyBudget(vo.getDailyBudget());
         entity.setBidStrategy(vo.getBidStrategy());
         entity.setOptPrice(vo.getOptPrice());
+        entity.setBidMultiplier(vo.getBidMultiplier());
         entity.setRemark(vo.getRemark());
         entity.setStatus(vo.getStatus());
         entity.setUpdateTime(new Date());
-        return updateById(entity);
+        AdGroupVO afterVO = Objects.requireNonNull(BeanUtil.copyProperties(entity, AdGroupVO.class));
+        bizLogApiService.logByUpdateAdGroupDirect(beforeVO, afterVO);
     }
 
     @Override
@@ -158,11 +185,15 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
         LambdaQueryWrapper<TargetCondition> wrapper = Wrappers.<TargetCondition>lambdaQuery()
                 .eq(TargetCondition::getAdGroupId, condition.getAdGroupId())
                 .eq(TargetCondition::getAttribute, ConditionEnum.BUNDLE.getDesc());
-        conditionService.remove(wrapper);
+        TargetCondition before = conditionService.getOne(wrapper);
+        if (before != null) {
+            conditionService.remove(wrapper);
+        }
         if (StrUtil.isNotBlank(condition.getValue())) {
             condition.setAttribute(ConditionEnum.BUNDLE.getDesc());
             conditionService.save(condition);
         }
+        bizLogApiService.logByUpdateAdGroupBundle(before, condition);
         return true;
     }
 
@@ -241,6 +272,8 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
 
     @Override
     public boolean updateBatch(BatchAdGroupUpdateVO vo) {
+        List<Integer> adGroupIds = vo.getAdGroupIds();
+        Map<Integer, AdGroup> beAdGroupMap = listByIds(adGroupIds).stream().collect(Collectors.toMap(IdEntity::getId, Function.identity()));
         List<AdGroup> adGroups = vo.getAdGroupIds().stream().map(id -> {
             AdGroup entity = new AdGroup();
             entity.setId(id);
@@ -252,6 +285,7 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
             return entity;
         }).collect(Collectors.toList());
         updateBatchById(adGroups);
+        bizLogApiService.logByUpdateBatch(beAdGroupMap, vo);
         return true;
     }
 
@@ -274,7 +308,11 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
         LambdaQueryWrapper<TargetCondition> wrapper = Wrappers.<TargetCondition>lambdaQuery()
                 .in(TargetCondition::getAdGroupId, adGroupIds)
                 .in(TargetCondition::getAttribute, ConditionEnum.BUNDLE.getDesc());
-        conditionService.remove(wrapper);
+        Map<Integer, TargetCondition> beConditonMap = conditionService.list(wrapper).stream()
+                .collect(Collectors.toMap(TargetCondition::getAdGroupId, Function.identity()));
+        if (!beConditonMap.isEmpty()) {
+            conditionService.remove(wrapper);
+        }
         if (StrUtil.isNotBlank(vo.getValue())) {
             List<TargetCondition> conditions = adGroupIds.stream().map(id -> {
                 TargetCondition condition = new TargetCondition();
@@ -286,6 +324,7 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
             }).collect(Collectors.toList());
             conditionService.saveBatch(conditions);
         }
+        bizLogApiService.logByUpdateBatchCondition("Bundle", beConditonMap, vo);
         return true;
     }
 
@@ -295,7 +334,11 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
         LambdaQueryWrapper<TargetCondition> wrapper = Wrappers.<TargetCondition>lambdaQuery()
                 .in(TargetCondition::getAdGroupId, adGroupIds)
                 .in(TargetCondition::getAttribute, ConditionEnum.HOUR.getDesc());
-        conditionService.remove(wrapper);
+        Map<Integer, TargetCondition> beConditonMap = conditionService.list(wrapper).stream()
+                .collect(Collectors.toMap(TargetCondition::getAdGroupId, Function.identity()));
+        if (!beConditonMap.isEmpty()) {
+            conditionService.remove(wrapper);
+        }
         if (StrUtil.isNotBlank(vo.getValue())) {
             List<TargetCondition> conditions = adGroupIds.stream().map(id -> {
                 TargetCondition condition = new TargetCondition();
@@ -307,6 +350,7 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
             }).collect(Collectors.toList());
             conditionService.saveBatch(conditions);
         }
+        bizLogApiService.logByUpdateBatchCondition("Hour", beConditonMap, vo);
         return true;
     }
 
@@ -317,7 +361,11 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
             LambdaQueryWrapper<TargetCondition> wrapper = Wrappers.<TargetCondition>lambdaQuery()
                     .in(TargetCondition::getAdGroupId, adGroupIds)
                     .in(TargetCondition::getAttribute, ConditionEnum.IMP_FREQUENCY.getDesc());
-            conditionService.remove(wrapper);
+            Map<Integer, TargetCondition> beConditonMap = conditionService.list(wrapper).stream()
+                    .collect(Collectors.toMap(TargetCondition::getAdGroupId, Function.identity()));
+            if (!beConditonMap.isEmpty()) {
+                conditionService.remove(wrapper);
+            }
             if (StrUtil.isNotBlank(vo.getImpValue())) {
                 List<TargetCondition> conditions = adGroupIds.stream().map(id -> {
                     TargetCondition condition = new TargetCondition();
@@ -329,12 +377,21 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
                 }).collect(Collectors.toList());
                 conditionService.saveBatch(conditions);
             }
+            BundleAdGroupUpdateVO bundleUpdateVO = new BundleAdGroupUpdateVO();
+            bundleUpdateVO.setAdGroupIds(adGroupIds);
+            bundleUpdateVO.setOperation(vo.getOperation());
+            bundleUpdateVO.setValue(vo.getImpValue());
+            bizLogApiService.logByUpdateBatchCondition("Imp Frequency", beConditonMap, bundleUpdateVO);
         }
         if (vo.getIsClickUpdate()) {
             LambdaQueryWrapper<TargetCondition> wrapper = Wrappers.<TargetCondition>lambdaQuery()
                     .in(TargetCondition::getAdGroupId, adGroupIds)
                     .in(TargetCondition::getAttribute, ConditionEnum.CLICK_FREQUENCY.getDesc());
-            conditionService.remove(wrapper);
+            Map<Integer, TargetCondition> beConditonMap = conditionService.list(wrapper).stream()
+                    .collect(Collectors.toMap(TargetCondition::getAdGroupId, Function.identity()));
+            if (!beConditonMap.isEmpty()) {
+                conditionService.remove(wrapper);
+            }
             if (StrUtil.isNotBlank(vo.getClickValue())) {
                 List<TargetCondition> conditions = adGroupIds.stream().map(id -> {
                     TargetCondition condition = new TargetCondition();
@@ -346,6 +403,11 @@ public class AdGroupServiceImpl extends ServiceImpl<AdGroupMapper, AdGroup> impl
                 }).collect(Collectors.toList());
                 conditionService.saveBatch(conditions);
             }
+            BundleAdGroupUpdateVO bundleUpdateVO = new BundleAdGroupUpdateVO();
+            bundleUpdateVO.setAdGroupIds(adGroupIds);
+            bundleUpdateVO.setOperation(vo.getOperation());
+            bundleUpdateVO.setValue(vo.getClickValue());
+            bizLogApiService.logByUpdateBatchCondition("Click Frequency", beConditonMap, bundleUpdateVO);
         }
         return true;
     }
